:sparkles: Implemented Registration :recycle: Cleaned up authentication service and utils..
