+++ cli/cmd/register/register.go
@@ -17,20 +17,20 @@ var RegisterCmd = &cobra.Command{
 	Long: "Register for the IHP CLI via keycloak",
 	Run: func(cmd *cobra.Command, args []string) {
 		if err := authentication.Register(); err != nil {
-			fmt.Println("Something went wrong while registering")
+			fmt.Println("Something went wrong while registering...")
 			fmt.Println(err)
 			os.Exit(1)
 		}
 
-		fmt.Println("Register successful!")
-		auth, err := authentication.GetSessionToken()
+		_, err := authentication.GetSessionToken()
 		if err != nil {
-			fmt.Println("Something went wrong while getting the session token after register")
-			fmt.Println("Please try to login again")
+			fmt.Println("Something went wrong while getting the session token after registration...")
+			fmt.Println("Please try again!")
 			fmt.Println(err)
 			os.Exit(1)
 		}
 
-		fmt.Println("Session token: " + auth)
+		fmt.Println("Registration successful! You can now use the IHP-CLI!.")
+		os.Exit(0)
 	},
 }
